requirements.txt
kafka-python
prometheus_client
python-json-logger
flask
waitress==2.1.2
pympler
requests